+++ kubernetes-tests/src/test/java/io/fabric8/kubernetes/client/mock/ErrorMessageTest.java
@@ -17,19 +17,16 @@ package io.fabric8.kubernetes.client.mock;
 
 import io.fabric8.kubernetes.api.model.Event;
 import io.fabric8.kubernetes.api.model.EventBuilder;
-import io.fabric8.kubernetes.api.model.EventList;
 import io.fabric8.kubernetes.api.model.Status;
 import io.fabric8.kubernetes.api.model.StatusBuilder;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.kubernetes.client.KubernetesClientException;
-import io.fabric8.kubernetes.client.dsl.NonNamespaceOperation;
 import io.fabric8.kubernetes.client.dsl.Resource;
 import io.fabric8.kubernetes.client.server.mock.EnableKubernetesMockClient;
 import io.fabric8.kubernetes.client.server.mock.KubernetesMockServer;
 import org.junit.jupiter.api.Test;
 
 import static java.net.HttpURLConnection.HTTP_BAD_REQUEST;
-import static java.net.HttpURLConnection.HTTP_FORBIDDEN;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatExceptionOfType;
 
 @EnableKubernetesMockClient
@@ -37,27 +34,6 @@ class ErrorMessageTest {
 
   KubernetesMockServer server;
   KubernetesClient client;
-
-  @Test
-  void whenCustomErrorMessageInConfig_thenErrorMessageShouldNotContainCustomMessage() {
-    // Given
-    String customMessage = "INVALID";
-    client.getConfiguration().getErrorMessages().put(403, customMessage);
-    server.expect().delete()
-        .withPath("/api/v1/namespaces/test/events")
-        .andReturn(HTTP_FORBIDDEN, new StatusBuilder()
-            .withCode(HTTP_FORBIDDEN)
-            .withMessage("forbidden")
-            .build())
-        .once();
-    NonNamespaceOperation<Event, EventList, Resource<Event>> eventResource = client.v1().events().inNamespace("test");
-
-    // When + Then
-    assertThatExceptionOfType(KubernetesClientException.class)
-        .isThrownBy(eventResource::delete)
-        .withMessageContaining(customMessage)
-        .hasFieldOrPropertyWithValue("code", HTTP_FORBIDDEN);
-  }
 
   @Test
   void whenResponseBodyContainsKubernetesStatus_thenErrorMessageShouldContainStatusMessage() {